book-recommendation-system-docker/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
DateTime==6.0
decorator==5.2.1
defusedxml==0.7.1
dnspython==2.8.0
docopt==0.6.2
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.133.1
fastjsonschema==2.21.2
filelock==3.20.0
fsspec==2025.12.0
gdown==5.2.1
greenlet==3.3.2
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pandocfilters==1.5.1
parso==0.8.6
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.9.2
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.2.3
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
webencodings==0.5.1
yarg==0.1.9
zope.interface==8.2
